webpack configuration, add components
