+++ .eslintrc.yml
@@ -3,4 +3,9 @@ env:
   browser: true
   commonjs: true
 
+parserOptions:
+    ecmaVersion: 2015
+    sourceType: 'module'
+
 extends: 'htmlacademy/es6'
+
