+++ webpack.config.js
@@ -1,0 +1,15 @@
+const path = require(`path`);
+module.exports = {
+  mode: `development`, // Режим сборки
+  entry: `./src/main.js`, // Точка входа приложения
+  output: { // Настройка выходного файла
+    filename: `bundle.js`,
+    path: path.join(__dirname, `public`)
+  },
+  devtool: `source-map`,
+  devServer: {
+    contentBase: path.join(__dirname, `public`),
+    publicPath: `http://localhost:8080`,
+    watchContentBase: true
+  }
+};
